move datastore to its own module file
