+++ packages/chrono-core/src/chrono.ts
@@ -1,29 +1,6 @@
 import { EventEmitter } from 'node:stream';
 
-export type TaskStatus = 'pending' | 'claimed' | 'completed' | 'failed';
-
-export type Task<TaskKind, TaskData> = {
-  id: string;
-  kind: TaskKind;
-  status: TaskStatus;
-  data: TaskData;
-  scheduledAt: Date;
-};
-
-type ScheduleInput<TaskKind, TaskData, DatastoreOptions> = {
-  when: Date;
-  kind: TaskKind;
-  data: TaskData;
-  priority?: 0;
-  idempotencyKey?: string;
-  datastoreOptions?: DatastoreOptions;
-};
-
-export interface Datastore<DatastoreOptions> {
-  schedule<TaskKind, TaskData>(
-    input: ScheduleInput<TaskKind, TaskData, DatastoreOptions>,
-  ): Promise<Task<TaskKind, TaskData>>;
-}
+import type { Datastore, ScheduleInput, Task } from './datastore';
 
 type ScheduleTaskInput<TaskKind, TaskData, DatastoreOptions> = ScheduleInput<TaskKind, TaskData, DatastoreOptions>;
